Upgraded to Gradle 7.1
