+++ .gitignore
@@ -29,3 +29,8 @@ out/
 # IDE files
 .idea/
 *.iml
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
