--- build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'java'
-    id 'jacoco'
-}
-
-group 'com.softwareonpurpose'
-sourceCompatibility = 11
-project.version = '1.0.9'
-
-repositories { mavenCentral() }
-
-dependencies {
-    implementation 'com.google.code.gson:gson:2.8.+'
-    testImplementation 'org.testng:testng:6.14.+'
-}
-
-wrapper { gradleVersion = '6.3' }
-
-jar { manifest { attributes 'Implementation-Title': 'indentmanager', 'Implementation-Version': archiveVersion } }
-
-test {
-    useTestNG()
-    systemProperties 'property': 'value'
-}
